requirements.txt
# Core
ultralytics>=8.4.0
mlflow>=2.10.0
pyyaml

# MLflow UI
flask

# Note: PyTorch is installed separately based on your hardware:
# CUDA (NYU server): pip install torch torchvision --index-url https://download.pytorch.org/whl/cu118
# MPS (MacBook):     pip install torch torchvision (default pip install works)
